==> 1. Data Preparation.ipynb <==
## The Data Preparation

The data is from [MovieTweetings Data](https://github.com/sidooms/MovieTweetings/tree/master/recsyschallenge2014).  To get started, you can read more about this project and the dataset from the [publication here](http://crowdrec2013.noahlab.com.hk/papers/crowdrec2013_Dooms.pdf).


import numpy as np
import pandas as pd

# Read in the MovieTweetings dataset originally taken from https://github.com/sidooms/MovieTweetings/tree/master/latest
movies = pd.read_csv('movies.dat', delimiter='::', header=None, names=['movie_id', 'movie', 'genre'], dtype={'movie_id': object}, engine='python')
reviews = pd.read_csv('ratings.dat', delimiter='::', header=None, names=['user_id', 'movie_id', 'rating', 'timestamp'], dtype={'movie_id': object, 'user_id': object, 'timestamp': object}, engine='python')


## For each of the datasets, there are a couple of cleaning steps we need to take care of:

### Movies
* Pulling the date from the title and creating new column
* Dummy the date column with 1's and 0's for each century of a movie (1800's, 1900's, and 2000's)
* Dummy column the genre with 1's and 0's

### Reviews
* Creating a date out of time stamp


def prepare_data(movies, reviews):
    '''
    A function for preparing and cleaning the movies and reviews dataframe performing the following tasks:
    - Pulling the date from the title and creating new column
    - Dummy the date column with 1's and 0's for each century of a movie (1800's, 1900's, and 2000's)
    - Dummy column the genre with 1's and 0's
    - Creating a date out of time stamp in reviews
    @param: movies and reviews dataframe Not cleaned
    @return: cleaned movies and reviews dataframe
    '''
    # Pulling the date from the title and create new column
    movies['year'] = movies.movie.apply(lambda x: x[-5:-1]
                                        if x[-1] == ')' else np.nan)
    movies['date'] = movies.movie.apply(lambda x: x[-5:-1]
                                        if x[-1] == ')' else np.nan)

    
    def dates(x):
        '''
        A function that return 1 if date is a given century or 0 if else 
        '''
        if x[:2] == yr:
            return 1
        else:
            return 0

    # Loop for mapping the dumming function with respect to the three centuries    
    for yr in ['18', '19', '20']:
        movies[yr + "00's"] = movies.year.apply(dates)

        
    # empty list to be filled with genres    
    genres = []

    # a loop for updating the genres list with the splitted values from the genre column
    for i in movies.genre:
        try:
            genres.extend(i.split('|'))
        except AttributeError:
            pass

    # Removing duplicated from the list    
    genres = set(genres)

    def gens(x):
        '''
        A function for searching for a certain genre in the genre columns row by row
        if found return 1 else return 0
        '''
        try:
            if x.find(i) > -1:
                return 1
            else:
                return 0
        except AttributeError:
            pass
    
    # loop for mapping the gens function over the genres collected
    for i in genres:
        movies[i] = movies.genre.apply(gens)
        
    import datetime
    # Creating a date out of timestamp
    reviews['date'] = reviews.timestamp.apply(lambda x: datetime.datetime.fromtimestamp(int(x)))
    
    return movies, reviews

movies, reviews = prepare_data(movies, reviews)

reviews.head()

movies.head()

reviews.to_csv('./reviews_clean.csv')
movies.to_csv('./movies_clean.csv')
